Fix FieldReference's simplified path
Fix bug in FieldReference's simplified implementation that tries
to copy null rows from a child of the RowVector which can be
outside its size.

Test Plan:
Caught this via #6600 but since then I am unable to create a test case
that fails without this change. Possibly because the conditions that
allowed nulls to be processed inside the simplified path have changed
between rebases. Still keeping the changes since that the correct way to
copy rows.