Fix: Avoid calling resize in addNulls for vectors of all encodings not only flat. (#6958)
Summary:

bikramSingh91 is working on a refactor to optimize AddNulls, in the meanwhile this diff
fix an exisiting bug, the previous fix ( https://github.com/facebookincubator/velox/pull/6856)
only apply to flat row, this make it work for all encodings.

To explain the error
- say input is Row of size 3 wrapping in dict.
- we call ensure writable then resize we get
row: size (5)
child size: 3.
- then we add nulls only to row 4, because 3 is not selected.
in that case the result vector is not valid and accessing row number (3) will
be invalid.

Fix https://github.com/facebookincubator/velox/issues/6935

Differential Revision: D50062849