Fix potential segfault in HashTable's parallelJoinBuilder (#6946)
Summary:

We have seen segfaults in HashTable's parallelJoinBuilder in the threads doing the parallel table partitioning.

The root cause seems to be:
* We 1 by 1 allocate a RowPartitions object and then spin up a thread asynchronously that uses that object to do table partitioning.
* Constructing a RowPartitions involves what can be a significant memory allocation.
* If constructing one of the RowPartitions triggers the MemoryPool to OOM, the resulting exception causes
the stack to unwind, notably destructing any RowPartitions objects we've constructed up to this point.
* Since the table partitioning threads we've already spawned may be actively using those RowPartitions
objects, we may see segfaults.

The fix is to declare the vector of RowPartitions before the sync guard that waits for the async threads to
finish, this way they are only destructed once the threads are finished with them.  This also protects us
against any other exceptions in this code that might cause it to exit while the table partitioning or table
building threads (which are also vulnerable) are still running.

I also split the loops for constructing the RowPartitions objects and spawning the table partitioning threads,
as this would protect us from any other potential concurrency issues caused by OOMs while creating
RowPartitions objects.

Reviewed By: xiaoxmeng

Differential Revision: D50035529