Ensure AsciiInfo is updated during resize() & prepareForReuse().
Summary:
We are seeing rare crashes and not so rare ASAN crashes during
Vector copy() when trying to process SimpleVector's AsciiInfo.
It happens so during the Vector recycling we don't touch AsciiInfo at all
and the new vector might have AsciiInfo from the previous one.
When accessing AsciiInfo it might be smaller than the vector and we are
accessing memory outside of the buffer.

To fix this we do the following:
1. Clear AsciiInfo in SimpleVector<StringView>::prepareForReuse().
2. Call SimpleVector<StringView>::resizeIsAsciiIfNotEmpty() from FlatVector<T>::resize().
3. Check if AsciiInfo is at least size of the SimpleVector in SimpleVector::validate().

Differential Revision: D50051589

fbshipit-source-id: 7969033b159696942214cba86997745062a99006