Fix TEXTFILE reading when filter on partition key but not project out
Summary: TEXTFILE table is pushing down partition key filters (not project out) down to the worker (DWRF table is not doing that in contrary).  These columns are not in data columns so it will give us an error when we try to get their type and constructing subfields pruning for them.  To handle this, we need to exclude them from subfields pruning construction.

Differential Revision: D50284303