Add ability for TableScan::getOutput() to exit with no data. (#7056)
Summary:

Add ability for TableScan::getOutput() to exit with no data.
For this we check two conditions before taking the next split to work on:
- Has our Task requested a stop for some reason. This is like the Driver does.
- Have we've been running for too long w/o results? New query config property used.

This is done to avoid long loops inside the operator for corner case queries, which
have highly filtering scan and return no rows for many splits.
This allows operator to help Driver exist the thread then the Task has requested it
and also would eliminate false positives for stuck operator calls.

Differential Revision: D50283576