Fix BaseVector::create() for RowVectors (#6967)
Summary:
Fix bug in BaseVector::create() that ends up creating a malformed
RowVector with no nulls but children of size 0. When copy is
called on this vector, it tries to access the valid rows at those
indices but since they are not marked as null, copy() tries to
access them in the children. The fix here is that when create() is
called and the children of target size are allocated, we should not
resize them to zero.

Additionally, as a part of fixing this bug, I had to refactor
how projections are passed to output in Operator and
NestedLoopJoinProbe. Currently, it created a RowVector of the
required size that would allocate all children of that size
and then replace each with projections. This is redundant work,
so instead I changed it to gather all projections/children and
then create a RowVector with them.

Finally, a test needed to be fixed in VectorSaver.constantRow where it
was comparing the value vector of a null constant row vector. Since the
value vector of a null constant is not serialized, there is no need to
compare its equality in the test.

Found by https://github.com/facebookincubator/velox/issues/6600


Differential Revision: D50103645

Pulled By: bikramSingh91