Fix setNullField in DWRF struct reader to handle missing subfield in file and newly created result
Summary:
Currently if a leaf level nested subfield is missing in file, and we
cannot reuse the result, the `VectorPtr` is null at that child and we try to
access type and pool information out of it.  Fix this by getting the type and
pool from the parent vector instead.

Differential Revision: D50422967