Make RowType's contructor throw VeloxRuntimeError.
Summary:
Make RowType's contructor throw VeloxRuntimeError
instead of VeloxUserError on "Mismatch names/types sizes".
We are seeing this error in batch workload and it is not
readily reproducible, which makes it a Runtime error.
We also want to see more details on the names and types.
Finally, we want to see the stack and maybe save
Vector + Expression.

Differential Revision: D50472707