Fix reserve bytes in ValueList (#7087)
Summary:
ValueList is used in array_agg and map_agg and related. It is an append only container for non-contiguous aggregation state. It previously left up to 1K of space after the last inserted value, which is excessive if values are short and containers are many. Now, this leaves up to the amount appended of trailing space, clamped between 16 and 1024.

Decreases worth case space consumption of array and related aggs by over 10x when arrays are short and groups numerous.


Reviewed By: xiaoxmeng

Differential Revision: D50352396

Pulled By: oerling