Prepare struct column reader for parallel decoding (#7151)
Summary:

We'll parallelize this loop with an executor in a future diff, so we're moving it to the end of the method to make it easier to reason that this method can return while the tasks initiated (in the future) in the loop are still running. The caller of the entire tree will have to wait though.

Differential Revision: D50459051