Fix a race condition in memmory pool reclaim
MemoryReclaimer::reclaim doesn't hold the reader lock when visit the child
memory pool which can cause race condition if there is a concurrent update
to the child memory pool. This race caused reclaimFromCompletedAggregation
and reclaimFromCompletedOrderBy test flakiness.
This PR fix the issue by holding a reader lock while accessing the child memory
pools and verified with the previous flaky tests with 1000 runs in clion and meta
internal.