Avoid reading index stream before loading
Summary:
In `SelectiveStringDictionaryColumnReader`, we read the row group index
stream in constructor by mistake.  This caused failed reading when
`BufferedInput` is used (`CachedBufferedInput` would work but performance is
suboptimal).  The row group index is already ensured to be loaded in
`SelectiveStringDictionaryColumnReader::ensureInitialized` which is the right
time for it to happen (after loading).

Differential Revision: D50658787