Introduce VeloxPersistentUserError exception and VELOX_USER_ABORT macros (#7238)
Summary:

Not all user errors should be suppressed by try expression, some errors are not Velox runtime errors
but are not meant to  be suppressed by try expression.

Right now we do not have a way to express them this PR add VeloxPersistentUserError which
represent such errors. and VELOX_USER_ABORT.. that can be used to throw them.

Differential Revision: D50661338