Fix metadata computation for shared lambda expressions
Currently the distinct fields for lambda expression are updated only
during construction. However, this can be an issue if we want to
clear and recompute it which happens when the expression is marked
as a shared expression. This change fixes it by ensuring recomputing
correctly updates the distinct fields everytime.

Test Plan:
Added unit test