Fix flaky failedToReclaimFromHashJoinBuildersInNonReclaimableSection
The test becomes flaky because we have moved non-reclaimable section check
in operator memory reclaim method instead of each individual spillable operator.
Correspondingly, if the reclaim is called on the hash build operator which is currently
under non-reclaimable section, then the injected testvalue callback in hash build
operator won't be executed. The test only pass if the reclaim method is called on
the hash build operator which is not under non-reclaimable section.

This PR fixes the flakiness by checking nonreclaimable section stats instead.