Add file id to spill merge stream
The id to SpillMergeStream to help order the spilled file from a given
spill partition. The spilled file id starts from zero of the first spilled file.
The in-memory spill stream has a special max uint32_t value. This
helps identify the first spilled file used in un-spilling merge path for
special handling such as distinct aggregation used to tell if a output
group is distinct or not.