Use huge pages for containers only
If huge pages are used for all allocations that contain an aligned 2
MB range, the system time for rewiring the memory may become
prohibitive.  We therefore use huge pages only for large long-lived
containers that are accessed at random, that is AllocationPool and
hash table arrays.

Adds a flag to ContiguousAllocation to indicate if huge pages are used.