Optimize HashStringAllocator::allocateFromFreeLists (#7313)
Summary:

In the currently implementation, we traverse the free list to look for a specific size for allocation.  This can be time-consuming for some query patterns (5x slower than Java).  Since the number of bytes we support in `HashStringAllocator` free lists are not large, we can afford one size class for each different size.  This way there is no need to traverse the list, and result in 30x CPU time improvement (6x faster than Java) for the slow queries.

The memory cost for each `HashStringAllocator` is about 36 KB.  For a typical query with 2 aggregate nodes and 15 drivers, it's about 1 MB overhead.  Usually hash aggregation takes hundreds of MB so the overhead is insignificant.  The new method also improves memory efficiency, so the net memory usage is most likely even decreasing.

Also allow different `HashTable`s to share same `HashStringAllocator` so that the memory overhead is minimized even for non-Presto client project, which instantiates more than 100k hash tables in a single process.

Reviewed By: oerling

Differential Revision: D50781449