Defer and reduce decoding and copying of non-key columns at TopN::addInput.
At first, only decodes the key columns. While processing the input, only copies the key columns for the row that's added to priority queue to RowContainer(to be used for later comparison). After the whole batch of input is all compared, decodes and copies the passed input rows to RowContainer.