Optimize dictionary memoization in favor of memory (#7226)
Summary:

We recently noticed that this optimization in expression evaluation
eagerly holds onto the base of the dictionary vector and the cached
results in anticipation that the next input vector might have the
same base. This additional memory can be held up unnecessarily in
cases where the input cannot take advantage of it. Therefore, with
this change we ensure that we only cache results if we encounter
the same base at least twice. Additionally, a strong reference to
the base is only held once results are cached. This is to ensure
that the vector held is not modified and re-used in-place.

Reviewed By: kagamiori

Differential Revision: D50622335