Control Prefetching
Pre-opening files and loading metadata is generally useful to
parallelize off the Driver control path. This is a latency dominated
activity.  Loading large numbers of columns is not as clearly
desirable since this can increase memory pressure by a multiple of the
hot columns in a stripe, up to hundreds of MB per driver in the worst
case.  So, add a hook that will decide whether there is space for
prefetch right before allocating the memory for the cache or other
buffers where the prefetch will go.

Adjust CachedBufferedInput coalescing to cut by size and not number of
streams. There can be hundreds of small densely accessed streams that
should coalesce. Also, for infrequently loaded data, lower the max
coalesce because infrequent loads may not be correlated. These come
from conditionals more than from top level conjuncts.