Ensure RowVector::resize also resizes the children. (#7431)
Summary:

This is the final change to enable resizing row vectors and their children. Previously we added,

1. Row vector validation ( https://github.com/facebookincubator/velox/pull/6481) (To ensure Rows are well formed via the validate api).
2. Row vector unsafe resize api (https://github.com/facebookincubator/velox/pull/7198) (To migrate over the few places which cannot use the newer api for performance or other reasons).

With this diff we make RowVector::resize recursively resize all the children along with the container. We also check to ensure that the children are uniquely referenced if the size is being increased.

This change also reverts:
1. https://github.com/facebookincubator/velox/pull/7010
2. https://github.com/facebookincubator/velox/pull/7069

Original commit changeset: bfe158c2d460

Original Phabricator Diff: D49547057

Reviewed By: laithsakka, bikramSingh91

Differential Revision: D49611624