fast path for storing a string into HashStringAllocator (#7456)
Summary:
Making a stream and appending a string and finishing a write to HashStringAllocator takes a long time.  The common use case is an append where there is one free list item that spans the rest of the last arena of the HSA. This makes a fast path for allocating the head of this fre block without first unhooking it from the free list and then putting the tail back there.


Reviewed By: Yuhta

Differential Revision: D51086021

Pulled By: oerling