Add termination time to TaskStats
Summary:
Like the title says, store termination time, separately from execution end time, in the TaskStats.

The reason for needing termination time comes from Presto.  They currently use execution end time to determine when a task was last used and clean up the Task once it's terminated and after some threshold of time after the end time.  If a task has partitioned output, that output may not be fully consumed until some time after the execution end time.  Termination time is guaranteed to be after the partitioned output has been consumed (assuming the query is running successfully), and so it's a better reflection of how long ago the task was used.

Today we see that for some queries, the execution end time is long before the termination time, and so Presto will delete the task as soon as it's terminated.  This leads to issues with in flight status checks Presto makes against the Worker.

Differential Revision: D51140307