Skip spilling if all the partitions have been spilled
If memory arbitration has reclaimed memory from the hash build operator which
request memory reservation, and there is still lack of memory, then the hash build
operator will try spill. However, this is only not required if the memory arbitration
has enabled. This can lead to a check failure in HashBuild::runSpill which expects
not all the partitions have been spilled.
This PR adds a check if all the partitions have been spilled when memory reservation
fails for a hash build operator, and skip spilling if all the partitions have been spilled.