Fix build issues for RowContainerSortBenchmark
RowContainerSortBenchmark depends on Parquet reader by default, but
Parquet support is controled by VELOX_ENABLE_PARQUET and it's defaulted
to OFF. This caused build failures because it tries to include some
Arrow headers which is not installed if VELOX_ENABLE_PARQUET is off.
This commit guard this benchmark by checking VELOX_ENABLE_PARQUET first.