Update Arrow namespaces in Parquet reader
Recently Arrow lib was partially moved into Parquet writer. Some of the
Parquet reader files still depends on some Arrow classes which will be
removed in the near future. The namespaces are now moved from arrow::
to ::arrow because it is still depending on the third_party arrow_ep
installation and this commit fixes it to avoid build failures.