Add a way to override current time in tests (#7543)
Summary:

Adding a ScopedTestTime class that allows us to override the value returned by
getCurrentTimeMs/getCurrentTimeMicro in tests.

This will be useful for testing some of the times we set in Task for example by making them deterministic.

It only applies in debug mode so there's no risk of breaking or impacting performance in prod.

Differential Revision: D51267554