Rename NoStop null handling mode to NullAsValue (#7599)
Summary:

This is the first diff of a sequence of diffs implementing the proposed solution in
https://github.com/facebookincubator/velox/issues/7536#issuecomment-1812947077

When nullHandlingMode is kNullAsValue mode null is treated as a value such that
- null == null is true.
- null == value is false.
- if flags.equalsOnly=false, order of null determined by the flag nullsFirst.

This diff also exposed some the current usages of kNoStop that are wrong and probably need to be changed.
this is not complete set since some other usages might be using the default constructor.

Differential Revision: D51378394