Add and test optional coroutine support (#7662)
Summary:

Add and test new co_routine methods.

Coroutines aren't supported by all compilers (specially in the open source repo), so the support must be optional.

To make the review easier, in this diff I'm just duplicating the `next` and `skip` methods to `co_next` and `co_skip` methods, and transforming them into coroutines by:
- making them return a `folly::coro::Task<>`
- adding at least a `co_return` in those methods
- Calling `co_` versions of methods from those transformed methods and `co_await`ing them

So this review is just doing that, we don't need to review the logic, since it's existing logic.

In the next diff I'll test all this, and from then on I'll start, diff my diff, to modify the logic to take advantage of those methods being co routines. I'll, for example, remove the use of executor.add(), and only parallelize on coroutines, with coroutine methods.

Once (and if) we're able to use coroutines in all compilers, and make sure that there aren't  performance regressions for running the `co_` versions of the methods when single threaded, we can remove the non-coroutine versions of the methods.

For now I'll keep the executor barrier. When I start refactoring the coroutine methods I'll remove that executor usage and schedule coroutines in the executor instead.

Differential Revision: D51479750