Fix fuzzer option containerHasNulls meaning (#7667)
Summary:

containerHasNulls used to control whether top-level container values can be
null, and the container elements themselves, as the name suggests. With this
change, I removed the reference to the flag to callsites where nullRatio is
already zero (and hence this is meaningless), check that other callsites are
covered and work well with the new meaning, and implemented the new semantic
described above.
Added unit tests to ensure it works as expected.

Reviewed By: xiaoxmeng

Differential Revision: D51487492