Add driver thread debug context when enqueue a closed driver
We found some time the driver thread debug context is not set on server
crash if it is triggered by driver destruction throws when we enqueue a
closed driver.
This PR adds the driver thread debug context when we enqueue a driver
to make sure any exception case is covered.